src/candiapp/parser.py
"""
Resume Parser - Core parsing logic for extracting structured data from resumes.

Targets industry benchmarks:
- Processing speed: <1 second per resume
- Parsing accuracy: >90%
- Data fields: 200+ fields
"""

import re
import time
from pathlib import Path
from typing import Optional, Dict, Any, List
import logging

# PDF parsing
try:
    import PyPDF2
    import pdfplumber
except ImportError:
    PyPDF2 = None
    pdfplumber = None

# DOCX parsing
try:
    import docx
    import docx2txt
except ImportError:
    docx = None
    docx2txt = None

from .models import (
    Resume,
    ParsedData,
    ContactInfo,
    Experience,
    Education,
    Skill,
    Language,
    Certification,
    ExperienceLevel,
    EducationLevel,
)

logger = logging.getLogger(__name__)


class ResumeParser:
    """
    Main resume parser class.

    Extracts structured data from resumes in multiple formats (PDF, DOCX, TXT).
    Designed to meet industry benchmarks for speed and accuracy.
    """

    SUPPORTED_FORMATS = {".pdf", ".docx", ".doc", ".txt"}

    # Regex patterns for common fields
    EMAIL_PATTERN = r"\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b"
    PHONE_PATTERN = r"(\+?\d{1,3}[-.\s]?)?\(?\d{2,4}\)?[-.\s]?\d{3,4}[-.\s]?\d{3,4}"
    LINKEDIN_PATTERN = r"(?:https?://)?(?:www\.)?linkedin\.com/in/[\w-]+"
    GITHUB_PATTERN = r"(?:https?://)?(?:www\.)?github\.com/[\w-]+"

    # Education keywords
    EDUCATION_KEYWORDS = [
        "bachelor", "master", "phd", "doctorate", "mba", "b.sc", "m.sc",
        "university", "college", "institute", "degree", "diploma",
        "בוגר", "תואר", "אוניברסיטה", "מכללה", "תעודה"
    ]

    # Experience section headers
    EXPERIENCE_HEADERS = [
        "experience", "work experience", "employment", "work history",
        "professional experience", "career history",
        "ניסיון", "ניסיון תעסוקתי", "עבודה"
    ]

    # Skills section headers
    SKILLS_HEADERS = [
        "skills", "technical skills", "core competencies", "expertise",
        "technologies", "tools", "programming languages",
        "כישורים", "מיומנויות", "טכנולוגיות"
    ]

    def __init__(self, enable_nlp: bool = False):
        """
        Initialize the resume parser.

        Args:
            enable_nlp: Enable NLP-based parsing (requires spaCy models)
        """
        self.enable_nlp = enable_nlp
        self._nlp_model = None

        if enable_nlp:
            try:
                import spacy
                # Try to load English and Hebrew models
                try:
                    self._nlp_model = spacy.load("en_core_web_sm")
                except OSError:
                    logger.warning("spaCy model 'en_core_web_sm' not found. Install with: python -m spacy download en_core_web_sm")
            except ImportError:
                logger.warning("spaCy not installed. NLP features disabled.")

    def parse(self, file_path: str) -> Resume:
        """
        Parse a resume file and extract structured data.

        Args:
            file_path: Path to the resume file

        Returns:
            Resume object with parsed data

        Raises:
            ValueError: If file format is not supported
            FileNotFoundError: If file does not exist
        """
        start_time = time.time()

        path = Path(file_path)
        if not path.exists():
            raise FileNotFoundError(f"Resume file not found: {file_path}")

        file_type = path.suffix.lower()
        if file_type not in self.SUPPORTED_FORMATS:
            raise ValueError(f"Unsupported file format: {file_type}")

        # Create Resume object
        resume = Resume(
            file_path=str(path),
            file_type=file_type[1:],  # Remove the dot
            file_size=path.stat().st_size,
        )

        try:
            # Extract text based on file type
            raw_text = self._extract_text(path, file_type)

            # Parse the extracted text
            parsed_data = self._parse_text(raw_text)
            parsed_data.raw_text = raw_text

            resume.parsed_data = parsed_data
            resume.parsing_time = time.time() - start_time

            logger.info(f"Parsed resume in {resume.parsing_time:.3f}s, extracted {resume.get_field_count()} fields")

        except Exception as e:
            logger.error(f"Error parsing resume: {e}")
            resume.parsing_errors.append(str(e))
            resume.parsing_time = time.time() - start_time

        return resume

    def _extract_text(self, path: Path, file_type: str) -> str:
        """Extract raw text from resume file."""
        if file_type == ".pdf":
            return self._extract_pdf_text(path)
        elif file_type in {".docx", ".doc"}:
            return self._extract_docx_text(path)
        elif file_type == ".txt":
            return path.read_text(encoding="utf-8", errors="ignore")
        else:
            raise ValueError(f"Unsupported file type: {file_type}")

    def _extract_pdf_text(self, path: Path) -> str:
        """Extract text from PDF file."""
        if pdfplumber is None:
            raise ImportError("pdfplumber is required for PDF parsing. Install with: pip install pdfplumber")

        text_parts = []
        try:
            with pdfplumber.open(path) as pdf:
                for page in pdf.pages:
                    text = page.extract_text()
                    if text:
                        text_parts.append(text)
        except Exception as e:
            logger.warning(f"pdfplumber failed, trying PyPDF2: {e}")
            # Fallback to PyPDF2
            if PyPDF2:
                with open(path, "rb") as f:
                    reader = PyPDF2.PdfReader(f)
                    for page in reader.pages:
                        text = page.extract_text()
                        if text:
                            text_parts.append(text)

        return "\n".join(text_parts)

    def _extract_docx_text(self, path: Path) -> str:
        """Extract text from DOCX file."""
        if docx is None:
            raise ImportError("python-docx is required for DOCX parsing. Install with: pip install python-docx")

        try:
            doc = docx.Document(path)
            text_parts = [para.text for para in doc.paragraphs]
            return "\n".join(text_parts)
        except Exception as e:
            logger.warning(f"python-docx failed, trying docx2txt: {e}")
            # Fallback to docx2txt
            if docx2txt:
                return docx2txt.process(str(path))
            raise

    def _parse_text(self, text: str) -> ParsedData:
        """
        Parse raw text and extract structured data.

        This is the core parsing logic that extracts all fields.
        """
        parsed_data = ParsedData()

        # Extract contact information
        parsed_data.contact = self._extract_contact_info(text)

        # Extract name (basic heuristic - first line or near contact info)
        parsed_data.full_name = self._extract_name(text)

        # Extract sections
        parsed_data.summary = self._extract_summary(text)
        parsed_data.experiences = self._extract_experiences(text)
        parsed_data.education = self._extract_education(text)
        parsed_data.skills = self._extract_skills(text)
        parsed_data.languages = self._extract_languages(text)
        parsed_data.certifications = self._extract_certifications(text)

        # Calculate derived fields
        parsed_data.total_experience_years = self._calculate_total_experience(parsed_data.experiences)
        parsed_data.experience_level = self._determine_experience_level(parsed_data.total_experience_years)

        return parsed_data

    def _extract_contact_info(self, text: str) -> ContactInfo:
        """Extract contact information from text."""
        contact = ContactInfo()

        # Email
        emails = re.findall(self.EMAIL_PATTERN, text)
        if emails:
            contact.email = emails[0]

        # Phone
        phones = re.findall(self.PHONE_PATTERN, text)
        if phones:
            contact.phone = phones[0]

        # LinkedIn
        linkedin_matches = re.findall(self.LINKEDIN_PATTERN, text, re.IGNORECASE)
        if linkedin_matches:
            contact.linkedin = linkedin_matches[0]

        # GitHub
        github_matches = re.findall(self.GITHUB_PATTERN, text, re.IGNORECASE)
        if github_matches:
            contact.github = github_matches[0]

        return contact

    def _extract_name(self, text: str) -> Optional[str]:
        """Extract candidate name (basic implementation)."""
        # Simple heuristic: first non-empty line that's not too long
        lines = text.split("\n")
        for line in lines[:10]:  # Check first 10 lines
            line = line.strip()
            if line and len(line) < 50 and not re.search(self.EMAIL_PATTERN, line):
                # Basic check if it looks like a name
                words = line.split()
                if 2 <= len(words) <= 4:  # Typical name has 2-4 words
                    return line
        return None

    def _extract_summary(self, text: str) -> Optional[str]:
        """Extract professional summary/objective."""
        # Look for summary section
        summary_patterns = [
            r"(?i)(?:professional\s+)?summary[:\s]+(.+?)(?=\n\n|\n[A-Z]|\Z)",
            r"(?i)objective[:\s]+(.+?)(?=\n\n|\n[A-Z]|\Z)",
            r"(?i)profile[:\s]+(.+?)(?=\n\n|\n[A-Z]|\Z)",
        ]

        for pattern in summary_patterns:
            match = re.search(pattern, text, re.DOTALL)
            if match:
                return match.group(1).strip()

        return None

    def _extract_experiences(self, text: str) -> List[Experience]:
        """Extract work experiences (basic implementation)."""
        experiences = []
        # This is a simplified implementation
        # A production version would use more sophisticated NLP
        return experiences

    def _extract_education(self, text: str) -> List[Education]:
        """Extract education entries (basic implementation)."""
        education = []
        # This is a simplified implementation
        # A production version would use more sophisticated NLP
        return education

    def _extract_skills(self, text: str) -> List[Skill]:
        """Extract skills from text."""
        skills = []
        # This is a simplified implementation
        # A production version would use NLP and skill taxonomies
        return skills

    def _extract_languages(self, text: str) -> List[Language]:
        """Extract language proficiencies."""
        languages = []
        # Basic pattern matching for languages
        return languages

    def _extract_certifications(self, text: str) -> List[Certification]:
        """Extract certifications."""
        certifications = []
        return certifications

    def _calculate_total_experience(self, experiences: List[Experience]) -> Optional[float]:
        """Calculate total years of experience."""
        # Simplified calculation
        return None

    def _determine_experience_level(self, years: Optional[float]) -> Optional[ExperienceLevel]:
        """Determine experience level based on years."""
        if years is None:
            return None
        if years < 2:
            return ExperienceLevel.ENTRY
        elif years < 5:
            return ExperienceLevel.JUNIOR
        elif years < 8:
            return ExperienceLevel.MID
        elif years < 12:
            return ExperienceLevel.SENIOR
        elif years < 15:
            return ExperienceLevel.LEAD
        else:
            return ExperienceLevel.EXECUTIVE
