pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "candiapp"
version = "0.1.0"
description = "AI-Powered Resume Parsing & Screening System"
readme = "README.md"
authors = [
    {name = "CandiApp Team", email = "info@candiapp.dev"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Human Resources",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Natural Language :: English",
    "Natural Language :: Hebrew",
]
keywords = ["resume", "parsing", "ats", "nlp", "recruitment", "hr", "screening"]

dependencies = [
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.0.0",
    "docx2txt>=0.8",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "regex>=2023.10.0",
    "python-dateutil>=2.8.0",
    "phonenumbers>=8.13.0",
    "email-validator>=2.1.0",
    "validators>=0.22.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
]
transformers = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
]
database = [
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/candiapp/candiapp"
Documentation = "https://candiapp.readthedocs.io"
Repository = "https://github.com/candiapp/candiapp"
"Bug Tracker" = "https://github.com/candiapp/candiapp/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=candiapp --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/candiapp"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
